refactor(interface): move shared types to central interface package
- Move SanitizedAgentOptions and related types to @tarko/interface
- Remove duplicate type definitions from web-ui
- Update server to use shared types from interface package
- Ensure proper type reuse across packages

This eliminates code duplication and ensures consistent type
definitions between server and client packages.
